fix(toolkit): field-labels map resolver falls back to humanizeFieldPath
The `provideFieldLabels(map)` JSDoc promised that unmapped keys would fall
back to `humanizeFieldPath`, matching the behavior documented in
docs/COMPLEX_NESTED_FORMS.md and docs/WARNINGS_SUPPORT.md. The
implementation instead returned the raw field path, so users got
`address.postalCode` instead of `Address / Postal code` for any key they
didn't explicitly map.

Move `humanizeFieldPath` (and the internal prefix helper) into
`core/utilities/humanize-field-path.ts` so the provider can use it without
a `core -> headless` cycle. The public headless export is preserved via
re-export.
